+++ Code/LYRUIConstants.m
@@ -1,52 +1,52 @@
 //
-//  LSUIConstants.m
+//  LYRUIConstants.m
 //  LayerSample
 //
 //  Created by Kevin Coleman on 6/17/14.
 //  Copyright (c) 2014 Layer, Inc. All rights reserved.
 //
 
-UIColor *LSBlueColor()
+UIColor *LYRUIBlueColor()
 {
     return [UIColor colorWithRed:36.0f/255.0f green:166.0f/255.0f blue:225.0f/255.0f alpha:1.0];
 }
 
-UIColor *LSDarkGrayColor()
+UIColor *LYRUIDarkGrayColor()
 {
     return [UIColor colorWithRed:148.0/255.0 green:148.0/255.0 blue:148.0/255.0 alpha:1.0];
 }
 
-UIColor *LSGrayColor()
+UIColor *LYRUIGrayColor()
 {
     return [UIColor colorWithRed:200.0/255.0 green:200.0/255.0 blue:200.0/255.0 alpha:1.0];
 }
 
-UIColor *LSLighGrayColor()
+UIColor *LYRUILightGrayColor()
 {
     return [UIColor colorWithRed:240.0f/255.0f green:240.0f/255.0f blue:240.0f/255.0f alpha:1.0];
 }
 
-UIColor *LSAddressBarGray()
+UIColor *LYRUIAddressBarGray()
 {
     return [UIColor colorWithRed:248.0f/255.0f green:248.0f/255.0f blue:248.0f/255.0f alpha:1.0];
 }
 
-UIColor *LSRedColor()
+UIColor *LYRUIRedColor()
 {
     return [UIColor colorWithRed:240.0f/255.0f green:80.0f/255.0f blue:100.0f/255.0f alpha:1.0];
 }
 
-UIFont *LSLightFont(CGFloat size)
+UIFont *LYRUILightFont(CGFloat size)
 {
     return [UIFont fontWithName:@"HelveticaNeue-Light" size:size];
 }
 
-UIFont *LSMediumFont(CGFloat size)
+UIFont *LYRUIMediumFont(CGFloat size)
 {
     return [UIFont fontWithName:@"HelveticaNeue" size:size];
 }
 
-UIFont *LSBoldFont(CGFloat size)
+UIFont *LYRUIBoldFont(CGFloat size)
 {
     return [UIFont fontWithName:@"HelveticaNeue-Medium" size:size];
 }