+++ Code/LYRUIConversationTableViewCell.h
@@ -19,7 +19,7 @@
  @abstract Customization selectors for configuring cell appearance
  */
 @property (nonatomic) UIFont *conversationLabelFont UI_APPEARANCE_SELECTOR;
-@property (nonatomic) UIColor *conversationLableColor UI_APPEARANCE_SELECTOR;
+@property (nonatomic) UIColor *conversationLabelColor UI_APPEARANCE_SELECTOR;
 
 @property (nonatomic) UIFont *lastMessageTextFont UI_APPEARANCE_SELECTOR;
 @property (nonatomic) UIColor *lastMessageTextColor UI_APPEARANCE_SELECTOR;
@@ -29,6 +29,6 @@
 
 @property (nonatomic) UIColor *unreadMessageIndicatorBackgroundColor UI_APPEARANCE_SELECTOR;
 
-@property (nonatomic) UIColor *backgroundColor UI_APPEARANCE_SELECTOR;
+@property (nonatomic) UIColor *cellBackgroundColor UI_APPEARANCE_SELECTOR;
 
 @end