+++ Code/LYRUIOutgoingMessageCollectionViewCell.m
@@ -15,7 +15,9 @@
     self = [super initWithFrame:frame];
     if (self) {
         self.avatarImageView.hidden = YES;
-
+        self.messageTextColor = [UIColor whiteColor];
+        self.bubbleViewColor = LYRUIBlueColor();
+        
         [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.bubbleView
                                                                      attribute:NSLayoutAttributeRight
                                                                      relatedBy:NSLayoutRelationEqual